Rename '/wards/schedule-confirmation' to '/wards/book-a-visit-confirmation'
